webserver/requirements.txt
redis
flask
Pillow
numpy